Player To Player Collision Implemented

+++ game-source-code/GameUpdater.cpp
@@ -9,7 +9,7 @@ void GameUpdater::updatePlayerLaser(LaserCanon& _laserCanon1, LaserCanon& _laser
     auto laser1 = 1;
     auto laser2 = 2;
 
-    if(_laserCanon1.getLaser(laser1).isAlive()) {
+    if(_laserCanon1.getLaser(laser1).isAlive() && _laserCanon1.isAlive()) {
         if(_laserCanon1.getLaser(laser1).getEntityCoordinates().getYposition() == 0) {
             _laserCanon1.destroyLaser(laser1);
         }
@@ -17,7 +17,7 @@ void GameUpdater::updatePlayerLaser(LaserCanon& _laserCanon1, LaserCanon& _laser
     }
     _laserCanon1.update_Laser_position(laser1);
 
-    if(_laserCanon2.getLaser(laser2).isAlive()) {
+    if(_laserCanon2.getLaser(laser2).isAlive() && _laserCanon2.isAlive()) {
         if(_laserCanon2.getLaser(laser2).getEntityCoordinates().getYposition() == 395) {
             _laserCanon2.destroyLaser(laser2);
         }
